+++ modules/multiresvolume/rendering/tfbrickselector.cpp
@@ -105,7 +105,7 @@ void TfBrickSelector::selectBricks(int timestep, std::vector<int>& bricks) {
             // would generate twice as much streaming. Current number of streams of this spatial brick
             // is 2^nTemporalSplits over the whole time period.
 
-            int newStreams = std::pow(2, bs.nTemporalSplits);
+            int newStreams = static_cast<int>(std::pow(2, bs.nTemporalSplits));
             //std::cout << "preparing for " << newStreams << " new streams" << std::endl;
 
             // Refining this one more step would require the double amount of streams
@@ -139,7 +139,7 @@ void TfBrickSelector::selectBricks(int timestep, std::vector<int>& bricks) {
             // On average on the whole time period, splitting this spatial brick into eight spatial bricks
             // would generate eight times as much streaming. Current number of streams of this spatial brick
             // is 2^nTemporalStreams over the whole time period.
-            int newStreams = 7*std::pow(2, bs.nTemporalSplits);
+            int newStreams = 7 * static_cast<int>(std::pow(2, bs.nTemporalSplits));
 
             if (nStreamedBricks + newStreams > totalStreamingBudget) {
                 // Reached dead end (streaming budget would be exceeded)
@@ -207,7 +207,7 @@ void TfBrickSelector::selectBricks(int timestep, std::vector<int>& bricks) {
             temporalSplitQueue.pop_back();
 
             unsigned int brickIndex = bs.brickIndex;
-            int newStreams = std::pow(2, bs.nTemporalSplits);
+            int newStreams = static_cast<int>(std::pow(2, bs.nTemporalSplits));
             if (nStreamedBricks + newStreams > totalStreamingBudget) {
                 // The current best choice would make us exceed the streaming budget, try next instead.
                 deadEnds.push_back(bs);
@@ -282,7 +282,7 @@ float TfBrickSelector::temporalSplitPoints(unsigned int brickIndex) {
     //std::cout << "local temporal split minimum for brick " << brickIndex << std::endl;
     diff = -diff;
     }
-    return diff * 0.5;
+    return diff * 0.5f;
 }
 
 float TfBrickSelector::spatialSplitPoints(unsigned int brickIndex) {
@@ -310,7 +310,7 @@ float TfBrickSelector::spatialSplitPoints(unsigned int brickIndex) {
     diff = -diff;
     }
 
-    return diff * 0.125;
+    return diff * 0.125f;
 }
 
 float TfBrickSelector::splitPoints(unsigned int brickIndex, BrickSelection::SplitType& splitType) {
@@ -345,7 +345,7 @@ bool TfBrickSelector::calculateBrickErrors() {
         glm::vec4 nextRgba = tf->sample(offset + 1);
 
         float colorDifference = glm::distance(prevRgba, nextRgba);
-        float alpha = (prevRgba.w + nextRgba.w) * 0.5;
+        float alpha = (prevRgba.w + nextRgba.w) * 0.5f;
 
         gradients[offset] = colorDifference*alpha;
     }
@@ -362,7 +362,7 @@ bool TfBrickSelector::calculateBrickErrors() {
             const Histogram* histogram = _histogramManager->getHistogram(brickIndex);
             float error = 0;
             for (int i = 0; i < gradients.size(); i++) {
-                float x = (i + 0.5) / tfWidth;
+                float x = (i + 0.5f) / tfWidth;
                 float sample = histogram->interpolate(x);
                 assert(sample >= 0);
                 assert(gradients[i] >= 0);
